move script to homeassistant.scripts module
